定点数优化。

+++ Assets/Script/FixedPoint/fixedPoint/fp.cs
@@ -151,6 +151,6 @@ public struct fp : IEquatable<fp>, IComparable<fp>
         return $"raw:【{rawValue}】float:【{(float)this}】";
     }
 
-    public int To32Fp => (int)rawValue;
+    public long To32Fp => rawValue;
 }
 #endif
